remove bigfile directory only for rank=0

+++ mpytools/io.py
@@ -849,10 +849,11 @@ class BigFile(BaseFile):
 
         if overwrite:
             self.mpicomm.Barrier()
-            try:
-                shutil.rmtree(os.path.join(self.filename, self.group.strip('/')), ignore_errors=True)  # otherwise previous columns are kept
-            except FileNotFoundError:
-                pass
+            if self.mpicomm.rank == 0:
+                try:
+                    shutil.rmtree(os.path.join(self.filename, self.group.strip('/')), ignore_errors=True)  # otherwise previous columns are kept
+                except FileNotFoundError:
+                    pass
             self.mpicomm.Barrier()
 
         with bigfile.FileMPI(comm=self.mpicomm, filename=self.filename, create=True) as file:
